changed dollar_amount type from float to decimal. updated needed_item model updated_funds to change received dollar_amount to decimal

+++ app/models/needed_item.rb
@@ -12,7 +12,7 @@ class NeededItem < ApplicationRecord
 
     def self.update_funds(shelter_id, item_id, dollar_amount)
         item = NeededItem.where(:shelter_id => shelter_id, :item_id => item_id).first
-        updated_funds = item.dollar_amount -= dollar_amount.to_f
+        updated_funds = item.dollar_amount -= dollar_amount.to_d
         item.update(dollar_amount: updated_funds)
     end
